ml-pipeline/model_training/requirements.txt
tensorflow==2.12.0
google-cloud-aiplatform==1.26.0
google-cloud-storage==2.8.0
pydicom==2.3.1
scikit-learn==1.2.2
pandas==2.0.2
numpy==1.23.5
google-cloud-monitoring==2.15.0